first iteration of update
